refactor: nesting folderでのコロケーション
