+++ src/application/web/routes/signup/route.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import type { MetaFunction } from '@remix-run/cloudflare';
-import useSignup from '../features/authenticate/useSignup';
-import SignupPage from '../features/authenticate/signup.page';
+import SignupPage from './signup.page';
+import useSignup from './useSignup';
 
 export const meta: MetaFunction = () => [{ title: 'アカウント作成 | TrendDiary' }];
