+++ src/application/web/routes/signup/useSignup.ts
@@ -1,11 +1,12 @@
 import { useState } from 'react';
 import { useNavigate } from '@remix-run/react';
-import { Errors, validateForm } from './validation';
+
 import getApiClientForClient from '../../infrastructure/api';
+import { AuthenticateErrors, validateAuthenticateForm } from '../../validation/authenticateForm';
 
 export default function useSignup() {
   const navigate = useNavigate();
-  const [errors, setErrors] = useState<Errors>({});
+  const [errors, setErrors] = useState<AuthenticateErrors>({});
   const [isLoading, setIsLoading] = useState(false);
 
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
@@ -14,7 +15,7 @@ export default function useSignup() {
     setErrors({});
 
     const formData = new FormData(e.currentTarget);
-    const validation = validateForm(formData);
+    const validation = validateAuthenticateForm(formData);
 
     if (!validation.isValid) {
       setErrors(validation.errors);
